WIP: implemented a different to handle requests.
Introduced Handler.accept(Incoming) so that Handlers can decide whether to accept the incoming request.
Processing of the incoming request is done via Incoming.accept(Processor).

This allows to split the API and avoid to give to application code a full fledged request API (e.g. ability to read the content) before the request is accepted.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>