WIP: merged Incoming into Request.
The main reason to have a single class is that there is no confusion about wrapping.
With Incoming, applications had to wrap Incoming and then wrap again Request, with the same logic used for Incoming, and this is prone to errors.

This change will expose methods such as Request.accept() and Request.readContent() that now should throw IllegalStateException if called at the wrong time, but it seems a better compromise.

Signed-off-by: Simone Bordet <simone.bordet@gmail.com>